+++ neurocaps/analysis/standardize.py
@@ -10,7 +10,7 @@ def standardize(subject_timeseries_list: Union[list[dict[str, dict[str, np.ndarr
                 file_names: Optional[list[str]]=None) -> dict[str, dict[str, dict[str, np.ndarray]]]:
 
     """
-    **Perform Participant-wise Timeseries Standardization**
+    Perform Participant-wise Timeseries Standardization.
 
     Standardizes the columns/ROIs of each run independently for all subjects in the subject timeseries. This function
     uses sample standard deviation, meaning Bessel's correction, `n-1` is used in the denominator. Note,